+++ CHANGELOG.md
@@ -5,12 +5,13 @@ All notable changes qsf-integration project will be documented in this file.
 ### Added
 - ElasticSearch Query Builder select the source fields automatically if they are mapped
 - add IOUtils to support io operations
+- add a unzip method to the IOUtils
+- implement a model loader
 
 ### Changed
 
-
 ### Fixed
-
+- close the http coneection in async elastic client
 
 ### Breaking Changes